Made Local Settings reporting generic
And a whole bunch of other stuff. Mainly, the descison was made so that
local settings, instead of reporting on a per-value basis, whole be
reported as a single property as a whole. Once Sturct were supported,
there were too many edge cases where users could define a property of a
struct, then a value inside of that sturct property as 2 different
columns that would compete when values were set.

Cleaned up the commerce references processors & mentions of sharepoint
in the code as well.

Consolidated the CSV constant Spring/Java definitions under CSV commons.

Deleted a bunch of unused classes & code